requirements.txt
joblib==1.2.0
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0.post5
threadpoolctl==3.1.0
